Fix: CVEs weren't shown after adding an API key
If you saw a stack trace fingerprint without having an API key
configured, therefore not seeing CVEs (if any), then configured an API
key and triggered the stack trace again, the extension would find that
it already did this API request and not request it again. Rather, after
adding an API key, it should retry the request, such that CVEs can be
found.